daaad_old/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.3
oauthlib==3.2.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.7.6
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sklearn-som==1.1.0
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
